reduce heartbeat error log
